requirements.txt
Flask==0.12.2
flask-restful==0.3.6
markdown==2.6.11
google-cloud-vision